src/backend/python/api-read/src/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.31
asyncpg==0.29.0
motor==3.7.1
aio_pika
pika
python-dotenv==1.0.1
pymongo==4.9.1
